UNROLL funcionando

+++ dgemm-code/DGEMM.h
@@ -20,7 +20,7 @@
 
 void dgemm(const size_t, const double*, const double*, double*);
 void dgemm_AVX(const size_t, double*, double*, double*);
-void dgemm_AVX_unroll(const int, double*, double*, double*);
+void dgemm_AVX_unroll(const size_t, double*, double*, double*);
 
 void aloca_matrizes(const size_t size, double **, double **, double **);
 void inicializa_matrizes(const size_t size, double *, double *, double *);
